Updated target file names to use dashes rather than underscores for consistency. Added the new common scripts made up of the average calculation and display script from the original display_bw script. These have already been implemented in the Makefile
